requirements.txt
autograd==1.6.1
cma==3.2.2
joblib==1.3.1
matlab==0.1
matplotlib==3.7.3
numpy==1.26.0
optuna==3.3.0
pandas==2.1.0
pyrecorder==0.2.1
pytest==7.4.1
PyYAML==6.0.1
scikit_learn==1.2.2
scipy==1.11.2
setuptools==68.2.1
pysqlite3==0.5.2
